requirements.txt
altair==5.5.0
asgiref==3.8.1
attrs==25.1.0
babel==2.17.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et_xmlfile==2.0.0
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.5
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
narwhals==1.26.0
numpy==1.26.2
openpyxl==3.1.5
packaging==24.2
pandas==2.1.3
passlib==1.7.4
pillow==11.1.0
protobuf==5.29.3
psycopg2-binary==2.9.9
pyarrow==19.0.0
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytube==15.0.0
pytz==2025.1
referencing==0.36.2
reportlab==4.4.0
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.3.2
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
tenacity==9.0.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
whitenoise==6.9.0
